Initial import refactoring
