+++ src/app/cluster/cluster-details/edit-provider-settings/aws-provider-settings/aws-provider-settings.component.spec.ts
@@ -13,10 +13,11 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 import {MatDialogRef} from '@angular/material/dialog';
 import {BrowserModule} from '@angular/platform-browser';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+
+import {ClusterService} from '@core/services';
+import {SharedModule} from '@shared/shared.module';
 import {ClusterMockService} from '@app/testing/services/cluster-mock-service';
 import {MatDialogRefMock} from '@app/testing/services/mat-dialog-ref-mock';
-import {ClusterService} from '@core/services/cluster/cluster.service';
-import {SharedModule} from '@shared/shared.module';
 import {AlibabaProviderSettingsComponent} from '../alibaba-provider-settings/alibaba-provider-settings.component';
 import {AzureProviderSettingsComponent} from '../azure-provider-settings/azure-provider-settings.component';
 import {DigitaloceanProviderSettingsComponent} from '../digitalocean-provider-settings/digitalocean-provider-settings.component';
@@ -27,6 +28,7 @@ import {KubevirtProviderSettingsComponent} from '../kubevirt-provider-settings/k
 import {OpenstackProviderSettingsComponent} from '../openstack-provider-settings/openstack-provider-settings.component';
 import {PacketProviderSettingsComponent} from '../packet-provider-settings/packet-provider-settings.component';
 import {VSphereProviderSettingsComponent} from '../vsphere-provider-settings/vsphere-provider-settings.component';
+
 import {AWSProviderSettingsComponent} from './aws-provider-settings.component';
 
 const modules: any[] = [BrowserModule, BrowserAnimationsModule, SharedModule];
