+++ src/app/cluster/cluster-details/edit-sshkeys/add-cluster-sshkeys/add-cluster-sshkeys.component.spec.ts
@@ -13,15 +13,19 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {MatDialog, MatDialogRef} from '@angular/material/dialog';
 import {BrowserModule} from '@angular/platform-browser';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {AppConfigService} from '../../../../config.service';
-import {ApiService, ClusterService, NotificationService, ProjectService, UserService} from '../../../../core/services';
-import {SharedModule} from '../../../../shared/shared.module';
-import {ApiMockService} from '../../../../testing/services/api-mock.service';
-import {AppConfigMockService} from '../../../../testing/services/app-config-mock.service';
-import {ClusterMockService} from '../../../../testing/services/cluster-mock-service';
-import {MatDialogRefMock} from '../../../../testing/services/mat-dialog-ref-mock';
-import {ProjectMockService} from '../../../../testing/services/project-mock.service';
-import {UserMockService} from '../../../../testing/services/user-mock.service';
+import {AppConfigService} from '@app/config.service';
+import {ApiMockService} from '@app/testing/services/api-mock.service';
+import {AppConfigMockService} from '@app/testing/services/app-config-mock.service';
+import {ClusterMockService} from '@app/testing/services/cluster-mock-service';
+import {MatDialogRefMock} from '@app/testing/services/mat-dialog-ref-mock';
+import {ProjectMockService} from '@app/testing/services/project-mock.service';
+import {UserMockService} from '@app/testing/services/user-mock.service';
+import {ApiService} from '@core/services/api/api.service';
+import {ClusterService} from '@core/services/cluster/cluster.service';
+import {NotificationService} from '@core/services/notification/notification.service';
+import {ProjectService} from '@core/services/project/project.service';
+import {UserService} from '@core/services/user/user.service';
+import {SharedModule} from '@shared/shared.module';
 import {AddClusterSSHKeysComponent} from './add-cluster-sshkeys.component';
 
 const modules: any[] = [BrowserModule, BrowserAnimationsModule, SharedModule];
