+++ src/app/core/services/cluster/cluster.service.ts
@@ -15,20 +15,20 @@ import {MatDialog, MatDialogConfig} from '@angular/material/dialog';
 import {combineLatest, merge, Observable, of, Subject, timer} from 'rxjs';
 import {catchError, filter, map, shareReplay, startWith, switchMap, switchMapTo, take} from 'rxjs/operators';
 
-import {environment} from '../../../../environments/environment';
-import {AppConfigService} from '../../../config.service';
-import {ConfirmationDialogComponent} from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
-import {LabelFormComponent} from '../../../shared/components/label-form/label-form.component';
-import {TaintFormComponent} from '../../../shared/components/taint-form/taint-form.component';
-import {Addon} from '../../../shared/entity/addon';
-import {Cluster, ClusterPatch, Finalizer, MasterVersion, ProviderSettingsPatch} from '../../../shared/entity/cluster';
-import {Event} from '../../../shared/entity/event';
-import {Health} from '../../../shared/entity/health';
-import {ClusterMetrics, NodeMetrics} from '../../../shared/entity/metrics';
-import {Node} from '../../../shared/entity/node';
-import {SSHKey} from '../../../shared/entity/ssh-key';
-import {CreateClusterModel} from '../../../shared/model/CreateClusterModel';
-import {ExternalClusterModel} from '../../../shared/model/ExternalClusterModel';
+import {environment} from 'environments/environment';
+import {AppConfigService} from '@app/config.service';
+import {ConfirmationDialogComponent} from '@shared/components/confirmation-dialog/confirmation-dialog.component';
+import {LabelFormComponent} from '@shared/components/label-form/label-form.component';
+import {TaintFormComponent} from '@shared/components/taint-form/taint-form.component';
+import {Addon} from '@shared/entity/addon';
+import {Cluster, ClusterPatch, Finalizer, MasterVersion, ProviderSettingsPatch} from '@shared/entity/cluster';
+import {Event} from '@shared/entity/event';
+import {Health} from '@shared/entity/health';
+import {ClusterMetrics, NodeMetrics} from '@shared/entity/metrics';
+import {Node} from '@shared/entity/node';
+import {SSHKey} from '@shared/entity/ssh-key';
+import {CreateClusterModel} from '@shared/model/CreateClusterModel';
+import {ExternalClusterModel} from '@shared/model/ExternalClusterModel';
 
 @Injectable()
 export class ClusterService {
